+++ inst/shiny/global.R
@@ -1,8 +1,24 @@
-library(shiny)
 library(polmineR)
 library(data.table)
-library(magrittr)
 loadNamespace("DT")
+
+check_and_install_pkg <- function(pkgname){
+  if (isFALSE(pkgname %in% rownames(installed.packages()))){
+    message(sprintf("The '%s' package is required to use the polmineR shiny app, ", pkgname),
+            "but it is not yet installed. Do you want to install it now?")
+    answer <- utils::menu(choices = c("Yes", "No"))
+    if (answer == 1L){
+      install.packages(pkgname)
+    } else {
+      stop(
+        sprintf("Aborting - the polmineR shiny app requires that the %s package is installed.", pkgname)
+      )
+    }
+  }
+}
+check_and_install_pkg("shiny")
+check_and_install_pkg("shinythemes")
+library(shiny)
 library(shinythemes)
 
 options(polmineR.shiny = TRUE)